client/telegram/requirements.txt
python-dotenv
aiogram
asyncio
aioredis
aioschedule
